+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to the package will be documented in this file.
 
+## [v6.8.1](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.8.1) - 2024-11-13
+
+- Fix CustomerAddressParams make all optional to avoid errors for partial data
+
+## [v6.8.0](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.8.0) - 2024-11-12
+
+- Update 20i to send additional customer data when creating stack users
+  - Attempt to update newly created stack users with name, address and phone number (in EPP format)
+
+## [v6.7.11](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.7.11) - 2024-11-11
+
+- Update Enhance findPlan() and findServerGroup() to match name case-insensitively
+
+## [v6.7.10](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.7.10) - 2024-10-15
+
+- Update Plesk
+  - Add API request/response debug logging
+  - Ignore errors fetching NS when DNS service is not enabled during getInfo()
+
 ## [v6.7.9](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.7.9) - 2024-09-30
 
 - Update 20i getLoginUrl() to always return a HTTPS link
@@ -47,6 +66,25 @@ All notable changes to the package will be documented in this file.
 - Update for PHP 8.1
 - Update base lib to v4
 - Update 20i SDK to v1.1
+
+## [v6.6.15](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.6.15) - 2024-11-13
+
+- Fix CustomerAddressParams make all optional to avoid errors for partial data
+
+## [v6.6.14](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.6.14) - 2024-11-12
+
+- Update 20i to send additional customer data when creating stack users
+  - Attempt to update newly created stack users with name, address and phone number (in EPP format)
+
+## [v6.6.13](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.6.13) - 2024-11-11
+
+- Update Enhance findPlan() and findServerGroup() to match name case-insensitively
+
+## [v6.6.12](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.6.12) - 2024-10-15
+
+- Update Plesk
+  - Add API request/response debug logging
+  - Ignore errors fetching NS when DNS service is not enabled during getInfo()
 
 ## [v6.6.11](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.6.11) - 2024-09-30